Utility functions now only in macros
